% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2.R
\name{pls2}
\alias{pls2}
\title{pls2 is an intern function of easyPLSDA package used to compute the PLS matrices using state of the art algorithms}
\usage{
pls2(X, Y, ncomp = 2, method = "classic", tol = 10^-9)
}
\arguments{
\item{X}{A matrix containing the explanatory variables of the model.}

\item{Y}{A dummy matrix of the response classes corresponding to the X observations.}

\item{ncomp}{The required number of components to compute.}

\item{method}{The method that will be used to compute the different matrices used in regression.
Default "classic" is for PLS2 algorithm, "SIMPLS" is the other method based on matrix SVD.}

\item{tol}{The convergence threshold for latent scores computation, only for "classic" method. Default: 10^-9.}
}
\value{
PLS2Reg S3 object.
\item{scores}{Scores matrices of X and Y variables projected into the new space}
\item{weights}{Matrices of weights for both X and Y variables}
\item{loadings}{One (or two for SIMPLS method) matrix of loadings}
\item{B.mat}{Only for "classic" mode. Matrix of "b" coefficients, result of the scalar product between X latent vectors and Y latent vectors}
\item{mode}{The mode used to compute the matrices}
}
\description{
pls2 is an intern function of easyPLSDA package used to compute the PLS matrices using state of the art algorithms
}
