% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsDA.R
\name{plsDA}
\alias{plsDA}
\title{plsDA function is an internal function of package easyPLSDA used to perform PLS-DA.}
\usage{
plsDA(X, Y, ncomp = 2, method = "classic", auto.select.var = FALSE,
  threshold = 0.8, tol = 10^-9)
}
\arguments{
\item{X}{A matrix containing the explanatory variables of the model.}

\item{Y}{A dummy matrix of the response classes corresponding to the X observations.}

\item{ncomp}{The required number of components to compute.}

\item{method}{The method that will be used to compute the different matrices used in regression.
Default "classic" is for PLS2 algorithm, "SIMPLS" is the other method based on matrix SVD.}

\item{auto.select.var}{Should the function automatically select variables using VIP model ? Default: True.}

\item{threshold}{The threshold for selecting the important variable.
Variables with weighted VIP under this threshold will be removed if auto.select.var is TRUE. Default: 0.8.}

\item{tol}{The convergence threshold for latent scores computation, only for "classic" method. Default: 10^-9.}
}
\value{
S3 object of class plsDA
\item{selected.var}{Names of explanatory variables selected}
\item{X}{Matrix of explanatory variables}
\item{scores}{Scores matrices of X and Y variables projected into the new space}
\item{weights}{Matrices of weights for both X and Y variables}
\item{loadings}{One (or two for SIMPLS method) matrix of loadings}
\item{mode}{The mode used to compute the matrices}
\item{explained.var}{The explained variance by latent vectors of X and Y}
\item{VIP}{Matrix of Variable Importance in Projection}
}
\description{
plsDA function is an internal function of package easyPLSDA used to perform PLS-DA.
}
