% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.PLSDA}
\alias{predict.PLSDA}
\title{Prediction method for PLSDA object}
\usage{
\method{predict}{PLSDA}(pls2Object, newdata, scale = TRUE)
}
\arguments{
\item{pls2Object}{Object of class PLSDA.}

\item{newdata}{data.frame or matrix, New observations for prediction.}

\item{scale}{default : TRUE, if the data has to be scaled before prediction computation.}
}
\value{
S3 object of class PLSDAprediction
}
\description{
Prediction method for PLSDA object
}
