% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PLSDA}
\alias{summary.PLSDA}
\title{Summary function for PLSDA objects}
\usage{
\method{summary}{PLSDA}(objectPLSDA)
}
\arguments{
\item{objectPLSDA}{}
}
\value{
Nothing
}
\description{
Summary function for PLSDA objects
}
