% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PLSDAprediction}
\alias{summary.PLSDAprediction}
\title{Summary function for PLSDAprediction objects}
\usage{
\method{summary}{PLSDAprediction}(object)
}
\arguments{
\item{object}{}
}
\value{
Nothing
}
\description{
Summary function for PLSDAprediction objects
}
